requirements.txt
# AutoDev - Core Dependencies

# LangChain ecosystem
langchain==0.1.0
langchain-anthropic==0.1.0
langgraph==0.0.20

# LLM providers
anthropic==0.18.0

# Vector database
chromadb==0.4.22

# Code analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code quality
pylint==3.0.3
black==23.12.1
